add validations users

+++ src/modules/accounts/useCases/deleteUser/DeleteUserUseCase.ts
@@ -3,6 +3,11 @@ import { inject, injectable } from 'tsyringe';
 import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
 import { AppError } from '@shared/errors/AppError';
 
+interface IRequest {
+  id: string;
+  userSession: string;
+}
+
 @injectable()
 class DeleteUserUseCase {
   constructor(
@@ -10,11 +15,15 @@ class DeleteUserUseCase {
     private usersRepository: IUsersRepository,
   ) {}
 
-  async execute(id: string): Promise<void> {
+  async execute({ id, userSession }: IRequest): Promise<void> {
     const userExists = await this.usersRepository.findById(id);
 
     if (!userExists) {
       throw new AppError('User not found!', 404);
+    }
+
+    if (userSession === id) {
+      throw new AppError('Operation not permitted', 401);
     }
 
     await this.usersRepository.delete(id);
